# openapi.yaml
openapi: 3.1.0
info:
  title: Banana Chat API
  version: 1.1.1
  description: REST contract for Banana Chat. Requirements are defined in PRODUCT_SPEC.md.
servers:
  - url: http://127.0.0.1:8000/api/v1
security: [{bearerAuth: []}]
paths:
...
  /workspace:
    get: {operationId: getWorkspace, parameters: [{$ref: '#/components/parameters/Workspace'}], responses: {'200': {description: Workspace detail}}}
...
  /devices:
    post: {operationId: registerDevice, parameters: [{$ref: '#/components/parameters/Workspace'}], responses: {'201': {description: Device registered}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
...
    Workspace: {name: X-Workspace-Id, in: header, required: true, schema: {type: string}}
...